test commit

+++ batch_2/print_remainder_of_two_numbers.py
@@ -1,6 +1,6 @@
 print("Please enter two numbers")
-num1 = float(input("1st number: "))
-num2 = float(input("2nd number: "))
+num1 = int(input("1st number: "))
+num2 = int(input("2nd number: "))
 if num2 == 0:
     print("Error: Division by zero is not allowed.")
 else:
